Improve decimal formatting in OSD + menu items
Summary:
- Replace decimal formatting specifiers with '%@' in i18n strings & use
  Double.string property extension in most places to ensure proper
  localized decimal & grouping separator punctuation.
- This commit extends to audio & subtitles (& seek menu items) what
  was already changed for speed-related items.
- Removal of rounding/truncation allows for more precise number display.

+++ iina/MainMenuActions.swift
@@ -302,8 +302,8 @@ extension MainMenuActionHandler {
   }
 
   @objc func menuChangeVolume(_ sender: NSMenuItem) {
-    if let volumeDelta = sender.representedObject as? Int {
-      let newVolume = Double(volumeDelta) + player.info.volume
+    if let volumeDelta = sender.representedObject as? Double {
+      let newVolume = volumeDelta + player.info.volume
       player.setVolume(newVolume)
     } else {
       Logger.log("sender.representedObject is not int in menuChangeVolume()", level: .error)
